# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BHRM Teams API
  version: 2.0.0
  description: |
    Enterprise Task-Centric Execution Platform — REST API.

    ## Authentication
    All protected routes require `Authorization: Bearer <token>` obtained from
    `POST /auth/login` or `POST /auth/refresh`.

    ## Response envelopes
    | Type | Shape |
    |---|---|
    | Single resource | `{ "success": true, "data": { … } }` |
    | Collection | `{ "success": true, "data": [ … ], "meta": { total, page, limit, totalPages } }` |
    | Error | `{ "success": false, "error": { "code": "…", "message": "…", "statusCode": 4xx } }` |

    ## Role model
    | Role | Description |
    |---|---|
    | `admin` | Company owner — full access to all company resources |
    | `staff` | Invited member — access scoped to assigned tasks |

    ## Endpoint naming conventions
    - Resource-based URLs, nouns only
    - State transitions via `PATCH /{resource}/{id}/status`
    - Bulk assignment via `PATCH /{resource}/{id}/assignees`
    - Avoid custom action verbs in URLs
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [admin, staff]
...
    ActivityAction:
      type: string
      enum: [task_created, task_updated, task_deleted, task_status_changed, task_assigned, task_unassigned, comment_added, comment_edited, comment_deleted, note_created, note_updated, note_deleted, attachment_uploaded, attachment_deleted, step_checked, step_unchecked, sprint_started, sprint_completed, project_archived, staff_invited, staff_suspended, staff_activated]

    # ── Pagination ────────────────────────────────────────────────────────────
    PaginatedMeta:
      type: object
      properties:
        total:      { type: integer, example: 42 }
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        totalPages: { type: integer, example: 3 }

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code:       { type: string, example: NOT_FOUND }
            message:    { type: string, example: Resource not found }
            statusCode: { type: integer, example: 404 }

    # ── User ─────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        fullName:  { type: string, example: Jane Smith }
        email:     { type: string, format: email }
        avatarUrl: { type: string, format: uri, nullable: true }
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          properties:
            role:            { $ref: '#/components/schemas/UserRole' }
            companyId:       { type: string, format: uuid, nullable: true }
            isEmailVerified: { type: boolean }
            lastActiveAt:    { type: string, format: date-time, nullable: true }
            createdAt:       { type: string, format: date-time }
            updatedAt:       { type: string, format: date-time }

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    AuthResponse:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        user:         { $ref: '#/components/schemas/UserProfile' }

    # ── Company ───────────────────────────────────────────────────────────────
...
    Attachment:
      type: object
      properties:
        id:          { type: string, format: uuid }
        taskId:      { type: string, format: uuid }
        fileName:    { type: string, example: design-spec.pdf }
        fileSize:    { type: integer, description: 'Bytes; max 10 485 760 (10 MB)' }
        mimeType:    { type: string, example: application/pdf }
        description: { type: string, nullable: true }
        uploader:    { $ref: '#/components/schemas/UserSummary' }
        downloadUrl: { type: string, format: uri }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }

    # ── Note ─────────────────────────────────────────────────────────────────
...
    CommentReply:
      type: object
      properties:
        id:        { type: string, format: uuid }
        content:   { type: string }
        author:    { $ref: '#/components/schemas/UserSummary', nullable: true }
        createdAt: { type: string, format: date-time }
    CommentReaction:
      type: object
      properties:
        id:    { type: string, format: uuid }
        emoji: { type: string, example: '👍' }
        user:  { $ref: '#/components/schemas/UserSummary' }
    CommentMention:
      type: object
      properties:
        userId:   { type: string, format: uuid }
        fullName: { type: string }
        email:    { type: string, format: email }
    Comment:
      type: object
      properties:
        id:      { type: string, format: uuid }
        taskId:  { type: string, format: uuid }
        content:
          type: string
          description: 'Content of the comment; "[deleted]" if soft-deleted'
        author:
          $ref: '#/components/schemas/UserSummary'
          nullable: true
        mentions:
          type: array
          items: { $ref: '#/components/schemas/CommentMention' }
        replies:
          type: array
          items: { $ref: '#/components/schemas/CommentReply' }
        reactions:
          type: array
          items: { $ref: '#/components/schemas/CommentReaction' }
        isEdited:  { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    # ── Activity ──────────────────────────────────────────────────────────────
    ActivityItem:
      type: object
      properties:
        id:         { type: string, format: uuid }
        action:     { $ref: '#/components/schemas/ActivityAction' }
        entityType:
          type: string
          enum: [task, sprint, project, comment, note, attachment, step, staff, company]
        entityId:   { type: string, format: uuid }
        entityName: { type: string, nullable: true }
        actor:      { $ref: '#/components/schemas/UserSummary', nullable: true }
        metadata:   { type: object, additionalProperties: true, nullable: true }
        createdAt:  { type: string, format: date-time }

    # ── Notification ──────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role or resource ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Duplicate record or constraint violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    MessageOk:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean }
              data:
                type: object
                properties:
                  message: { type: string }

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
...
    taskId:       { name: taskId,       in: path,  required: true, schema: { type: string, format: uuid } }
...
    commentId:    { name: commentId,    in: path,  required: true, schema: { type: string, format: uuid } }
    replyId:      { name: replyId,      in: path,  required: true, schema: { type: string, format: uuid } }
...
    attachmentId: { name: attachmentId, in: path,  required: true, schema: { type: string, format: uuid } }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new admin account and company
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName, companyName]
              properties:
                email:       { type: string, format: email }
                password:    { type: string, minLength: 8, description: 'Min 8 chars, ≥1 uppercase, ≥1 digit' }
                fullName:    { type: string }
                companyName: { type: string }
      responses:
        '201':
          description: Account and company created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/AuthResponse' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive access + refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password-reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Reset email sent (always 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      message: { type: string }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Set a new password using the reset token from email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:    { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '400': { description: Token expired or invalid }
  /auth/accept-invite:
    post:
      tags: [Auth]
      summary: Accept a staff invitation and activate account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, fullName, password]
              properties:
                token:    { type: string }
                fullName: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Invite accepted — tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/AuthResponse' }
        '400': { description: Token expired or already used }
        '404': { $ref: '#/components/responses/NotFound' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Verify token and return current user
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserProfile' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ════════════════════════════════════════════════════════════════════════════
  # USERS
  # ════════════════════════════════════════════════════════════════════════════
...
  /tasks/{taskId}/audit:
    get:
      tags: [Tasks]
      summary: Audit log for a task — all field changes with before/after values (admin only)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ActivityItem' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tasks/{taskId}/attachments:
    get:
      tags: [Attachments]
      summary: List attachments for a task
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Attachment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Attachment' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Attachments]
      summary: Upload a file (assignee or admin; max 10 MB; MIME validated)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
                  description: Optional description saved as metadata
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Attachment' }
        '400': { description: File exceeds 10 MB }
        '403': { $ref: '#/components/responses/Forbidden' }
        '415': { description: MIME type not allowed }
  /tasks/{taskId}/attachments/{attachmentId}:
    patch:
      tags: [Attachments]
      summary: Update attachment metadata — description, fileName (admin or uploader)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/attachmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
                fileName:    { type: string, description: 'Display name only; physical file is unchanged' }
      responses:
        '200':
          description: Updated attachment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Attachment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Attachments]
      summary: Delete an attachment (admin always; staff only if they uploaded it)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/attachmentId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/attachments/{attachmentId}/download:
    get:
      tags: [Attachments]
      summary: Stream download a file attachment
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/attachmentId'
      responses:
        '200':
          description: Binary file stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════════════════════
  # NOTES
  # ════════════════════════════════════════════════════════════════════════════
...
  /tasks/{taskId}/comments:
    get:
      tags: [Comments]
      summary: List comments including replies and reactions (newest first)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/page'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Comment' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Comments]
      summary: Post a comment (assignee or admin; mentionedUserIds validated against active staff)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:          { type: string, minLength: 1, maxLength: 5000 }
                mentionedUserIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Comment posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Comment' }
        '400': { description: Invalid mention — user not active staff }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/{taskId}/comments/{commentId}:
    patch:
      tags: [Comments]
      summary: Edit a comment (admin always; staff only if they authored it)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/commentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:          { type: string, minLength: 1, maxLength: 5000 }
                mentionedUserIds: { type: array, items: { type: string, format: uuid } }
      responses:
        '200':
          description: Updated comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Comment' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Comments]
      summary: Delete a comment (admin always; staff only if they authored it — content becomes "[deleted]")
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/commentId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/comments/{commentId}/replies:
    post:
      tags: [Comments]
      summary: Post a reply to a comment (assignee or admin)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/commentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:          { type: string, minLength: 1, maxLength: 5000 }
                mentionedUserIds: { type: array, items: { type: string, format: uuid } }
      responses:
        '201':
          description: Reply posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/CommentReply' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/comments/{commentId}/replies/{replyId}:
    delete:
      tags: [Comments]
      summary: Delete a reply (admin always; staff only if they authored it)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/commentId'
        - $ref: '#/components/parameters/replyId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /tasks/{taskId}/comments/{commentId}/reactions:
    post:
      tags: [Comments]
      summary: Toggle an emoji reaction — posting the same emoji twice removes it
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/commentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji: { type: string, maxLength: 10, example: '👍' }
      responses:
        '200':
          description: Reaction toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      toggled:  { type: boolean, description: 'true = added; false = removed' }
                      reaction: { $ref: '#/components/schemas/CommentReaction', nullable: true }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════════════════════
  # DASHBOARD
  # ════════════════════════════════════════════════════════════════════════════
...